use bounded kinds to prevent parser errors
I faced issues while using ccls because it has additional kinds which
are not presented in the kinds array (see
https://github.com/MaskRay/ccls/blob/f8d2778b659a12a94b60258f949373ad42f964a1/src/lsp.hh#L173)
It triggers parser errors. To deal with this kind of problem, I added
the bounded_kind function, which limits the kind values and uses a
default value if the index is out of the band.